Add files via upload
I was able to write this code for Task 2 for this assignment, but when I print out the code it doesn't give the specific information in that node, just gibberish. The code runs but I'm not sure if the output is gonna count against me or if that output is what I'm supposed to get. Thanks Professor! Have a good weekend.